+++ CHANGELOG.md
@@ -6,6 +6,9 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [Unreleased]
+- placeholder
+
+## [0.2.0] - 2016-11-05
 
 ### Added
 - New environment variable CLIKRAKEN_DEFAULT_PAIR to override the default currency pair.
@@ -81,7 +84,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 ### Changed
 - API key file location moved to '~/.config/clikraken/kraken.key'
 
-[Unreleased]: https://github.com/zertrin/clikraken/compare/0.1.9...HEAD
+[Unreleased]: https://github.com/zertrin/clikraken/compare/0.2.0...HEAD
+[0.2.0]: https://github.com/zertrin/clikraken/compare/0.1.9...0.2.0
 [0.1.9]: https://github.com/zertrin/clikraken/compare/0.1.8...0.1.9
 [0.1.8]: https://github.com/zertrin/clikraken/compare/0.1.7...0.1.8
 [0.1.7]: https://github.com/zertrin/clikraken/compare/0.1.6...0.1.7